requirements.txt
selenium
python-dotenv
requests
smtplib
email